1.2.0+2025/06/28
Se Corrige el DataLoader, Se agregan comentarios con Swagger, y se empiezan a programar las pruebas unitarias de rolService (completado parcialmente) y credencialService (en procesos)

+++ src/main/java/com/SAFE_Rescue/API_Administrador/repository/BomberoRepository.java
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Repository;
  * Maneja validadores para encontrar el run y telefono
  */
 @Repository
-public interface BomberoRepository extends JpaRepository<Bombero , Long> {
+public interface BomberoRepository extends JpaRepository<Bombero , Integer> {
 
-    public boolean existsByRun(Integer run);
+    public boolean existsByRun(int run);
 
-    public boolean existsByTelefono(Integer telefono);
+    public boolean existsByTelefono(int telefono);
 
 }
